+++ newclient/src/modules/address/components/TableAddressTxsRow.vue
@@ -32,7 +32,7 @@
                     <v-flex xs6>
                         <v-layout row align-center justify-end>
                             <app-time-ago :timestamp="transfer.getTimestamp()" class="info--text caption" />
-                            <app-state-diff v-if="!isPending" :state="state" class="ml-2 mr-1" />
+                            <app-state-diff v-if="!isPending && !loadingStateDiff && transfer.stateDiff" :state="state" class="ml-2 mr-1" />
                             <p v-if="isMinedIndicator && isPending" class="caption primary--text blinking ml-2">{{ $t('tx.mined') }}</p>
                         </v-layout>
                     </v-flex>
@@ -69,7 +69,7 @@
                     =====================================================================================
                     -->
                     <v-flex xs7 sm9 pa-1>
-                        <app-transform-hash :hash="typeAddr | toChecksum" :link="`/address/${typeAddr}`" />
+                        <app-transform-hash v-if="!isContractCreation" :hash="typeAddr | toChecksum" :link="`/address/${typeAddr}`" />
                     </v-flex>
                 </v-layout>
             </div>
@@ -105,7 +105,12 @@
                                     </v-card>
                                 </v-flex>
                                 <v-flex sm7 lg8 pl-0>
-                                    <app-transform-hash :hash="typeAddr | toChecksum" :link="`/address/${typeAddr}`" :italic="true" />
+                                    <app-transform-hash
+                                        v-if="!isContractCreation"
+                                        :hash="typeAddr | toChecksum"
+                                        :link="`/address/${typeAddr}`"
+                                        :italic="true"
+                                    />
                                 </v-flex>
                             </v-layout>
                         </v-flex>
@@ -173,7 +178,7 @@
                     <v-layout row align-center justify-end>
                         <v-icon v-if="transfer.getStatus()" small class="txSuccess--text">fa fa-check-circle</v-icon>
                         <v-icon v-else small class="txFail--text">fa fa-times-circle</v-icon>
-                        <app-state-diff :state="state" class="ml-3 mr-1" />
+                        <app-state-diff v-if="!loadingStateDiff && transfer.stateDiff" :state="state" class="ml-3 mr-1" />
                     </v-layout>
                 </v-flex>
                 <v-flex v-else shrink>
@@ -196,7 +201,7 @@ import { FormattedNumber, NumberFormatHelper } from '@app/core/helper/number-for
 import { EthTransfer } from '@app/modules/address/models/EthTransfer'
 import BN from 'bignumber.js'
 
-const TYPES = ['in', 'out', 'self']
+const TYPES = ['in', 'out', 'self', 'contractCreation']
 @Component({
     components: {
         AppTooltip,
@@ -215,11 +220,18 @@ export default class TableTxsRow extends Mixins(NumberFormatMixin) {
     @Prop(String) address!: string
     @Prop({ type: Boolean, default: false }) isPending!: boolean
     @Prop(Boolean) isMinedIndicator?: boolean
+    @Prop(Function) getStateDiff!: (_hash: string, _type: string) => void
+    @Prop(Boolean) loadingStateDiff?: boolean
     /*
     ===================================================================================
-      Initial Data
+      Lifecycle
     ===================================================================================
     */
+    created() {
+        if (this.transfer.getStatus() === false && !this.transfer.stateDiff) {
+            this.getStateDiff(this.transfer.getHash(), this.type)
+        }
+    }
 
     /*
     ===================================================================================
@@ -227,23 +239,26 @@ export default class TableTxsRow extends Mixins(NumberFormatMixin) {
     ===================================================================================
     */
     get state(): object {
-        const stateData = [{ name: `${this.$t('state.bal-before')}`, value: this.transfer.getBalBefore(this.type) }]
-        if (this.type !== TYPES[0]) {
-            stateData.push({ name: `${this.$tc('tx.fee', 1)}`, value: this.transfer.getFee() })
+        if (!this.loadingStateDiff) {
+            const stateData = [{ name: `${this.$t('state.bal-before')}`, value: this.transfer.getBalBefore(this.type) }]
+            if (this.type !== TYPES[0]) {
+                stateData.push({ name: `${this.$tc('tx.fee', 1)}`, value: this.transfer.getFee() })
+            }
+            stateData.push({
+                name: this.getValueTitle,
+                value: this.transfer.getStatus() === false ? NumberFormatHelper.formatNonVariableEthValue(new BN(0)) : this.transfer.getValue()
+            })
+            return {
+                status: this.transfer.getStatus(),
+                balAfter: this.transfer.getBalAfter(this.type),
+                data: stateData
+            }
         }
-        stateData.push({
-            name: this.getValueTitle,
-            value: this.transfer.getStatus() === false ? NumberFormatHelper.formatNonVariableEthValue(new BN(0)) : this.transfer.getValue()
-        })
-        return {
-            status: this.transfer.getStatus(),
-            balAfter: this.transfer.getBalAfter(this.type),
-            data: stateData
-        }
+        return {}
     }
 
     get getValueTitle(): string {
-        if (this.type === TYPES[1]) {
+        if (this.type === TYPES[1] || this.type === TYPES[3]) {
             if (this.transfer.getStatus() === false) {
                 return `${this.$t('state.actual-sent')}`
             }
@@ -270,6 +285,9 @@ export default class TableTxsRow extends Mixins(NumberFormatMixin) {
     get type(): string {
         const from = this.transfer.getFrom().toLowerCase()
         const to = this.transfer.getTo().toLowerCase()
+        if (to === '') {
+            return TYPES[3]
+        }
         const addr = this.address.toLowerCase()
         if (addr === from && addr === to) {
             return TYPES[2]
@@ -284,6 +302,8 @@ export default class TableTxsRow extends Mixins(NumberFormatMixin) {
                 return `${this.$t('tx.type.in')}`
             case TYPES[1]:
                 return `${this.$t('tx.type.out')}`
+            case TYPES[3]:
+                return `${this.$t('contract.creation')}`
             default:
                 return `${this.$t('tx.type.self')}`
         }
@@ -294,6 +314,8 @@ export default class TableTxsRow extends Mixins(NumberFormatMixin) {
                 return 'primary'
             case TYPES[1]:
                 return 'error'
+            case TYPES[3]:
+                return 'warning'
             default:
                 return 'info'
         }
@@ -324,9 +346,6 @@ export default class TableTxsRow extends Mixins(NumberFormatMixin) {
     }
     get transferValueClass(): string {
         let textColor = 'black--text'
-        // if (this.isPending) {
-        //     console.log(this.transfer)
-        // }
         if (this.transfer.getStatus() === null) {
             textColor = 'info--text'
         }
@@ -340,6 +359,13 @@ export default class TableTxsRow extends Mixins(NumberFormatMixin) {
     }
     get hasFeeSign(): boolean {
         return this.transfer.getStatus() !== null && this.type === TYPES[1]
+    }
+    get isMined(): boolean {
+        return !this.transfer.getIsPending()
+    }
+
+    get isContractCreation(): boolean {
+        return this.type === TYPES[3]
     }
 }
 </script>